chore: modularize WalletGatewayApi
