+++ .swiftlint.yml
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: EUPL-1.2
 
 included:
+  - Packages
   - Wallet
   - WalletTests
   - WalletUITests
