+++ Packages/WalletGatewayApi/WalletGateway/Sources/WalletGateway/Internal/HTTPRequest+Extensions.swift
