+++ Packages/WalletGatewayApi/WalletGateway/Sources/WalletGateway/OpenAPI/openapi-generator-config.yaml
