+++ Packages/WalletGatewayApi/WalletGateway/Sources/WalletGateway/Session/SessionError.swift
