+++ Packages/WalletGatewayApi/WalletGateway/Sources/WalletGateway/Session/SessionManager.swift
