--- WalletGateway/Sources/WalletGateway/SecurityEnvelope.swift
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Digg - Agency for digital government
-//
-// SPDX-License-Identifier: EUPL-1.2
-
-public struct SecurityEnvelope: Sendable {
-  public let content: String
-}
